refactor: move components into src/components
