+++ src/components/Background.svelte
