+++ src/components/ErrorModal.svelte
